Split UnixDomainSocketTest netcoreapp specific pieces to only be compiled for netcoreapp.
Also - fix the new test names to follow the existing format in the file.
